+++ src/neuroconv/datainterfaces/ecephys/openephys/openephysbinarydatainterface.py
@@ -87,7 +87,7 @@ class OpenEphysBinaryRecordingInterface(BaseRecordingExtractorInterface):
                 "More than one stream is detected! Please specify which stream you wish to load with the `stream_name` argument. "
                 "To see what streams are available, call `OpenEphysRecordingInterface.get_stream_names(folder_path=...)`."
             )
-        if len(available_streams) > 1 and stream_name not in available_streams:
+        if stream_name is not None and stream_name not in available_streams:
             raise ValueError(
                 f"The selected stream '{stream_name}' is not in the available streams '{available_streams}'!"
             )